backend/pyproject.toml
[tool.poetry]
name = "dbt-dashboard"
version = "0.1.0"
description = ""
authors = ["Chris Stinson <chris.stinson@esr.cri.nz>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.3"
fastapi = "^0.115.6"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pytz = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
